Split IngestBatcherTest into two files
